requirements.txt
matplotlib
numpy
opencv-python
pillow
pyzbar
qrcode
fastapi
uvicorn
python-multipart
authlib
itsdangerous
httpx
jinja2
redis